services/pipeline/requirements.txt
ccxt==4.3.59
pandas==2.2.2
pyarrow==17.0.0
boto3==1.34.159
pydantic==2.8.2
loguru==0.7.2
python-dateutil==2.9.0.post0
requests==2.32.3
python-telegram-bot==21.4
textblob==0.18.0.post0
nltk==3.9.1
psycopg2-binary==2.9.9
statsmodels==0.14.2
matplotlib==3.9.2
openai==1.46.0
prometheus-client==0.20.0
sentry-sdk==2.13.0
redis==5.0.7
optuna==3.6.1
mlflow==2.16.0
feedparser==6.0.11
